Address feedback: refactor Duo views
- Created a BaseDuoForm that has all the common code, so now EnrollView and ChallengeView just specify the title and the footer
- Changed css selectors to use idx classnames
