+++ assets/sass/v2/_duo.scss
@@ -1,5 +1,5 @@
-.enroll-duo form,
-.mfa-verify-duo form {
+.enroll-authenticator--duo_native form,
+.challenge-authenticator--duo_native form {
   iframe {
     height: 400px;
     width: 100%;
